add manager
