+++ packages/editor/src/lib/editor/types/external-content.ts
@@ -2,7 +2,7 @@ import { VecLike } from '@tldraw/primitives'
 import { EmbedDefinition } from '@tldraw/tlschema'
 
 /** @public */
-export type TLCreateShapeFromSourceInfo =
+export type TLExternalContent =
 	| {
 			type: 'text'
 			point?: VecLike
