+++ packages/ui/src/lib/hooks/useInsertMedia.ts
@@ -14,7 +14,7 @@ export function useInsertMedia() {
 		async function onchange(e: Event) {
 			const fileList = (e.target as HTMLInputElement).files
 			if (!fileList || fileList.length === 0) return
-			await editor.onCreateShapeFromSource({
+			await editor.onPutExternalContent({
 				type: 'files',
 				files: Array.from(fileList),
 				point: editor.viewportPageBounds.center,
